feat(server): 不再区分Maintainer

+++ server/src/main/kotlin/api/RouteArticle.kt
@@ -8,7 +8,6 @@ import infra.article.Article
 import infra.comment.CommentRepository
 import infra.common.Page
 import infra.user.UserOutline
-import infra.user.UserRole
 import io.ktor.resources.*
 import io.ktor.server.plugins.ratelimit.*
 import io.ktor.server.request.*
@@ -193,7 +192,7 @@ class ArticleApi(
         validatePageNumber(page)
         validatePageSize(pageSize)
 
-        val ignoreHidden = user != null && user.role atLeast UserRole.Maintainer
+        val ignoreHidden = user != null && user.role atLeast UserRole.Admin
 
         return articleRepo
             .listArticle(
@@ -245,7 +244,7 @@ class ArticleApi(
         user: User?,
         id: String,
     ): ArticleDto {
-        val ignoreHidden = user != null && user.role atLeast UserRole.Maintainer
+        val ignoreHidden = user != null && user.role atLeast UserRole.Admin
 
         val article = articleRepo.getArticle(id)
             ?: throwArticleNotFound()
@@ -321,7 +320,7 @@ class ArticleApi(
         user: User,
         id: String,
     ) {
-        user.requireMaintainer()
+        user.requireAdmin()
         val isDeleted = articleRepo.deleteArticle(id = id)
         if (!isDeleted) throwArticleNotFound()
         commentRepo.deleteCommentBySite("article-${id}")
@@ -332,7 +331,7 @@ class ArticleApi(
         id: String,
         pinned: Boolean,
     ) {
-        user.requireMaintainer()
+        user.requireAdmin()
         val isUpdated = articleRepo.updateArticlePinned(id = id, pinned = pinned)
         if (!isUpdated) throwArticleNotFound()
     }
@@ -342,7 +341,7 @@ class ArticleApi(
         id: String,
         locked: Boolean,
     ) {
-        user.requireMaintainer()
+        user.requireAdmin()
         val isUpdated = articleRepo.updateArticleLocked(id = id, locked = locked)
         if (!isUpdated) throwArticleNotFound()
     }
@@ -352,7 +351,7 @@ class ArticleApi(
         id: String,
         hidden: Boolean,
     ) {
-        user.requireMaintainer()
+        user.requireAdmin()
         val isUpdated = articleRepo.updateArticleHidden(id = id, hidden = hidden)
         if (!isUpdated) throwArticleNotFound()
     }
